Renaming, extra final refactoring
