+++ PythonClient/carla/agent_benchmark/experiment_suite/sample.py
@@ -11,15 +11,17 @@ from __future__ import print_function
 import os
 import numpy as np
 
-from .benchmark import Benchmark
-from .experiment import Experiment
+from carla.agent_benchmark.benchmark import Benchmark
+from carla.agent_benchmark.experiment import Experiment
 from carla.sensor import Camera
 from carla.settings import CarlaSettings
 
-from .metrics import Metrics
+from carla.planner.planner import Planner
+
+from carla.agent_benchmark.metrics import Metrics
 
 
-class CoRL2017(Benchmark):
+class SampleBenchmark(Benchmark):
 
     def __init__(self,
                  city_name,
@@ -31,14 +33,15 @@ class CoRL2017(Benchmark):
 
         # Define all the parameters used to compute the driving summary.
 
+        # TODO: this should be taken from the default class
 
         metrics_parameters = {
 
-            'intersection_offroad': {'frames_skip': 10,
+            'intersection_offroad': {'frames_skip': 10,  # Check intersection always with 10 frames tolerance
                                      'frames_recount': 20,
                                      'threshold': 0.3
                                      },
-            'intersection_otherlane': {'frames_skip': 10,
+            'intersection_otherlane': {'frames_skip': 10,  # Check intersection always with 10 frames tolerance
                                        'frames_recount': 20,
                                        'threshold': 0.4
                                        },
@@ -57,15 +60,14 @@ class CoRL2017(Benchmark):
             'dynamic_episodes': [3]
 
         }
+
         self._metrics = Metrics(metrics_parameters)
 
         # All the weather used on this benchmark
-        self._weathers = [1, 3, 6, 8, 4, 14]
-        # Improve readability by adding a weather dictionary
-        self._weather_name_dict = {1: 'Clear Noon', 3: 'After Rain Noon',
-                                   6: 'Heavy Rain Noon', 8: 'Clear Sunset',
-                                   4: 'After Rain Sunset', 14: 'Soft Rain Sunset'}
+        self._weathers = [1]
 
+
+        # TODO: Simplify this initalization, Default parameters go to benchmark
         Benchmark.__init__(self, city_name,
                            name_to_save,
                            continue_experiment,
@@ -82,16 +84,11 @@ class CoRL2017(Benchmark):
             return len(self._poses_town02()[0])
 
 
-    def get_all_statistics(self):
-
-        summary = self._metrics.compute(self._recording._path)
-
-        return summary
 
 
 
 
-    def _calculate_time_out(self, start_point, end_point):
+    def _calculate_time_out(self, start_point, end_point): #TODO this could come from base class
         """
         Function to return the timeout ( in miliseconds) that is calculated based on distance to goal.
         This is the same timeout as used on the CoRL paper.
@@ -106,30 +103,18 @@ class CoRL2017(Benchmark):
 
     def _poses_town01(self):
         """
-        Each matrix is a new task. We have all the four tasks
+            Each matrix is a new task. We have all the four tasks
 
         """
 
         def _poses_straight():
-            return [[36, 40], [39, 35], [110, 114], [7, 3], [0, 4],
-                    [68, 50], [61, 59], [47, 64], [147, 90], [33, 87],
-                    [26, 19], [80, 76], [45, 49], [55, 44], [29, 107],
-                    [95, 104], [84, 34], [53, 67], [22, 17], [91, 148],
-                    [20, 107], [78, 70], [95, 102], [68, 44], [45, 69]]
+            return [[36, 40]]
 
         def _poses_one_curve():
-            return [[138, 17], [47, 16], [26, 9], [42, 49], [140, 124],
-                    [85, 98], [65, 133], [137, 51], [76, 66], [46, 39],
-                    [40, 60], [0, 29], [4, 129], [121, 140], [2, 129],
-                    [78, 44], [68, 85], [41, 102], [95, 70], [68, 129],
-                    [84, 69], [47, 79], [110, 15], [130, 17], [0, 17]]
+            return [[138, 17]]
 
         def _poses_navigation():
-            return [[105, 29], [27, 130], [102, 87], [132, 27], [24, 44],
-                    [96, 26], [34, 67], [28, 1], [140, 134], [105, 9],
-                    [148, 129], [65, 18], [21, 16], [147, 97], [42, 51],
-                    [30, 41], [18, 107], [69, 45], [102, 95], [18, 145],
-                    [111, 64], [79, 45], [84, 69], [73, 31], [37, 81]]
+            return [[105, 29]]
 
         return [_poses_straight(),
                 _poses_one_curve(),
@@ -139,26 +124,13 @@ class CoRL2017(Benchmark):
     def _poses_town02(self):
 
         def _poses_straight():
-            return [[38, 34], [4, 2], [12, 10], [62, 55], [43, 47],
-                    [64, 66], [78, 76], [59, 57], [61, 18], [35, 39],
-                    [12, 8], [0, 18], [75, 68], [54, 60], [45, 49],
-                    [46, 42], [53, 46], [80, 29], [65, 63], [0, 81],
-                    [54, 63], [51, 42], [16, 19], [17, 26], [77, 68]]
+            return [[38, 34]]
 
         def _poses_one_curve():
-            return [[37, 76], [8, 24], [60, 69], [38, 10], [21, 1],
-                    [58, 71], [74, 32], [44, 0], [71, 16], [14, 24],
-                    [34, 11], [43, 14], [75, 16], [80, 21], [3, 23],
-                    [75, 59], [50, 47], [11, 19], [77, 34], [79, 25],
-                    [40, 63], [58, 76], [79, 55], [16, 61], [27, 11]]
+            return [[37, 76]]
 
         def _poses_navigation():
-            return [[19, 66], [79, 14], [19, 57], [23, 1],
-                    [53, 76], [42, 13], [31, 71], [33, 5],
-                    [54, 30], [10, 61], [66, 3], [27, 12],
-                    [79, 19], [2, 29], [16, 14], [5, 57],
-                    [70, 73], [46, 67], [57, 50], [61, 49], [21, 12],
-                    [51, 81], [77, 68], [56, 65], [43, 54]]
+            return [[19, 66]]
 
         return [_poses_straight(),
                 _poses_one_curve(),
@@ -168,9 +140,8 @@ class CoRL2017(Benchmark):
 
     def _build_experiments(self):
         """
-        Creates the whole set of experiment objects,
-        The experiments created depend on the selected Town.
-
+            Creates the whole set of experiment objects,
+            The experiments created depends on the selected Town.
 
         """
 
@@ -228,9 +199,9 @@ class CoRL2017(Benchmark):
 
         return experiments_vector
 
-    def _get_details(self):
+    def _get_details(self): #TODO Automatize to get values directly from the master class.
 
         # Function to get automatic information from the experiment for writing purposes
-        return 'corl2017_' + self._city_name
+        return 'sample_' + self._city_name
